+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - `placeholder` property to input widget.
 - `transition_duration` property to stack widget.
 - `widget_control` utility function for dynamic widget handling.
+- `text` and `show_text` property to progressbar widget.
 
 ## [0.3.1] - 2025-11-01